# Task/Update log format on dev and prod