--- a/package.json
+++ b/package.json
@@ -17,11 +17,12 @@
     "dev": "nodemon --max-http-header-size=40000 ./bin/www",
     "lint": "node_modules/eslint/bin/eslint.js config routes bin/www app.js",
     "start": "NODE_ENV=production node --max-http-header-size=40000 ./bin/www",
+    "start:cypress": "NODE_ENV=test node ./bin/www",
     "test": "node node_modules/jest/bin/jest.js",
     "cypress:open": "./node_modules/.bin/cypress open",
-    "cypress": "start-server-and-test start http://localhost:3005 cypress:open",
+    "cypress": "start-server-and-test start:cypress http://localhost:3005 cypress:open",
     "cypress:run": "./node_modules/.bin/cypress run",
-    "cypress:cli": "start-server-and-test start http://localhost:3005 cypress:run"
+    "cypress:cli": "start-server-and-test start:cypress http://localhost:3005 cypress:run"
   },
   "dependencies": {
     "app-root-path": "^2.2.1",