--- a/package.json
+++ b/package.json
@@ -44,8 +44,7 @@
     "purecss": "^1.0.1",
     "request-promise": "^4.2.4",
     "trim-request": "^1.0.6",
-    "validator": "^11.1.0",
-    "winston": "^3.2.1"
+    "validator": "^11.1.0"
   },
   "devDependencies": {
     "axe-core": "^3.3.2",